pyproject.toml
[project]
name = "model-context-protocol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "langchain-groq>=0.3.2",
    "mcp-use>=1.2.6",
    "mcp[cli]>=1.6.0",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]
